requirements.txt
selenium
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==2.1.2
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.7
Markdown==3.0.1
meld3==1.0.2
pycparser==2.19
pytz==2018.5
six==1.11.0
psycopg2-binary